Keep reference to EnchantBroker when Dicts are live
Modelled after the C++ bindings, the library assumed that the `EnchantBroker` will live for the whole lifetime of the `EnchantDict`s created by it. This is not a safe assumption to make.

If the `EnchantBroker` were freed but the `EnchantDict` were still live, it would produce a warning. And what is worse, the Dict’s broker field would then point to dead memory, potentially causing issues during memory consolidation.

To make it safe, we started wrapping the EnchantBroker pointer in Rc.

This solution is ugly and unidiomatic – it would likely be better to utilize Rust’s ownership system – but I am not that familiar with Rust and this workaround is easy enough.

Works around: https://github.com/patronus-checker/enchant-rs/issues/4